Refactors config path
