+++ components/addToWatchlist.js
@@ -1,4 +1,4 @@
-const { movieCollection, userCollection, ObjectId } = require("../config");
+const { movieCollection, userCollection, ObjectId } = require("../setup/config.js");
 
 // Add movie to the user's watchlist
 async function addToWatchlist(req, res) {
