+++ components/resetPasswordToken.js
@@ -1,4 +1,4 @@
-const { userCollection } = require("../config");
+const { userCollection } = require("../setup/config.js");
 
 async function resetPasswordToken(req, res) {
   const { token } = req.params;
